Introduce Store#findRecordByKey
Used to find records by key instead of local id.

First looks up a local id for a given key and, if none is present in the
keyMap, generates a new local id and adds it to the keyMap.

Next `findRecord` is called for the `type` and local `id`.

In this way the same id:key mapping will be accessible for other 
sources.