Ensure that Store#addRecord always returns a record on success
In blocking coordination scenarios, the store’s `beforeUpdate` event may be used to `sync` a change back to the store before `update` is called. In this scenario, `addRecord` will not return a new record.

Instead, we should prepare the record instance before calling `update`. This instance can then be used to look up the model from the identity map regardless of the return value of `update`.