[BREAKING] Configurable lookup paths; change default for models (`models/` -> `data-models/`)
This commit adds support for configuring how ember-orbit will lookup the modules
it uses and how it injects them into your app. It is configured under an orbit entry in
config/environment.js. The default values are:

```
orbit: {
  types: {
    model: 'data-model',
    source: 'data-source',
    strategy: 'data-strategy'
  },
  collections: {
    models: 'data-models',
    sources: 'data-sources',
    strategies: 'data-strategies'
  },
  services: {
    store: 'store',
    coordinator: 'data-coordinator',
    schema: 'data-schema',
    keyMap: 'data-key-map'
  }
}
```

 - `types` controls naming in ember's container.
 - `collections` controls naming in ember's module/filesystem. Currently,
   collections should be the plural version of the corresponding type
 - `services` controls naming of the services that ember-orbit provides
   and the properties they are injected onto

Setting `orbit.types.model` to `data-model` and `orbit.collections.models`
to `data-models` means that you would put your orbit model classes in files within
`app/data-models/`. This is a breaking change from 0.15, where they were looked up within
`app/models/`. This change was made for consistency and to make ember-orbit better able to
co-exist with ember-data in an app. In order to restore the 0.15 behavior, use the following
config:

```
orbit: {
  types: {
    model: 'model'
  },
  collections: {
    models: 'models'
  }
}
```

The `services` config nodes determines what name is used when injecting
the ember-orbit store service onto routes and controllers. If you want
to avoid a conflict with ember-data's injections, change this value. e.g.

```
orbit: {
  services: {
    store: 'orbit-store'
  }
}
```

Note that service names are camelized when injecting, so a value of `orbit-store`
would be injected as `orbitStore`