Only restrict mutations on model fields for non-forked stores
This refactor more narrowly confines the concept of "mutable models" to mean direct mutation of model fields. For a store that is not a fork, direct sync mutation of model fields will be prevented, but async mutation through methods like `replaceAttribute` or `update` will still be allowed.

This prevents the common pitfalls associated with setting `planet.name = 'somethingInvalid';` - namely that changes aren't reflected immediately in records and any async errors won't be caught. However, it still allows `await planet.replaceAttribute('name', 'somethingInvalid')`, which is less ergonomic but more robust. If direct sync mutation is needed, say for two-way-bound inputs in a form, it's recommended that a fork be created which will still allow this approach.