Prefix all standard methods and properties on Model class with `$`
As discussed in #283, Orbit's `Model` class has long had the potential for namespace conflicts, since the built-in methods and properties are co-mingled with any custom field properties that access the model's underlying data.

This deprecates all the built-in methods and properties on `Model` and replaces them with versions that are prefixed with `$`.

Thus, for a `Model` instance `planet`, replace calls to `planet.remove()` with `planet.$remove()`. And instead of accessing `planet.store`, access `planet.$store`. The properties `id` and `type` are not deprecated, as they access underlying record data. However, `identity` is an ember-orbit construct and has been deprecated in favor of `$identity`.

Closes #283