Allow undefined responses from update / merge
Some coordination scenarios can result in `update` or `merge` (which calls `update` internally) to return `undefined`. For instance, if a blocking `beforeUpdate` listener applies the transform back to the source via `sync`, then `update` itself will return `undefined` because that transform will already have been logged.

Because of this, we need to check that a data response is not `undefined` before choosing to lookup results in their model form.