Narrow Model.$isDisconnected tracking
Instead of directly tracking a model's `_cache`, track a new protected property, `_isDisconnected`, in order to determine when a model has been disconnected from its cache.

This fixes a problem, reported in https://github.com/orbitjs/ember-orbit/issues/359, in which rendered static resultsets would immediately be invalidated when a member was removed.

Note that, with this change, it's still possible to make a model's rendering conditional upon its `$isDisconnected` property.