Expose $store on Model
The recent changes (thanks!) to cache change tracking replaced `Model`'s `$store` with `$cache`.

But `$cache` is not fully-featured in the same way that `$store` is. For example, it doesn't implement `fork` and `merge`. `$cache.sourceCache` does implement these methods, but at the orbit layer, not the ember-orbit layer, meaning we don't get back a new `Cache` that can give us `Model` instances.

This PR re-exposes `$store` on `Model` (and also `store` on `Cache`). This makes is again possible to take an existing Model and fork it on demand, and then merge back into its original source.

An alternative solution here would be to implement fork & merge on ember-orbit's `Cache`. But I'm not 100% clear if those are intended to have the same semantics as forking and merging of underlying sources. In my use case, I definitely want the merge to not only merge into the base cache, but merge all the way into the base source.

The thing I'm trying to keep working is this pair of utility methods that are used throughout my app to accept an existing model and establish a fork of it that can later by merged:

```js
export function fork(model) {
  let store = model.$store;
  let forkedStore = store.fork();
  return forkedStore.cache.findRecord(model);
}

export async function merge(model) {
  let { $store } = model;
  if (!$store.forked) {
    throw new Error(`tried to save a model that is not in a forked store`);
  }
  let result = await $store.base.merge($store);
  return result[0];
}
```