Change `store` to be optional on `Cache`
This will allow for forked caches to exist that are not directly associated with a `Store`.

To reach the `store` from a Model, you will need to access it via `modelInstance.$cache.store` to keep the relationship chain clear and so that typings are clear that a cache's `store` may be `undefined` (vs. a model must have a `$cache`).