Introduce main bucket service to allow for bucket injections
By registering a bucket with a standard service name, the `applyStandardSourceInjections` utility can inject it into sources. The standard bucket service name can be configured as an orbit option, just like the other standard orbit services.

A note has been added to the bucket initializer blueprint to clarify that only one bucket should be registered as the standard bucket service.

This eliminates the last usage of implicit injections in ember-orbit.

It's a breaking change because previously created bucket initializers should be revisited to remove implicit injections and replace them with the service registration.