Failing test for remoteId on unloaded relationship with data
In https://github.com/orbitjs/ember-orbit/issues/400 @dgeb suggested that even unloaded records should have a type, id, and (if configured) remoteId key when accessed via a relationship that loaded with a `data` property.

These tests demonstrate that that works when the related record is already loaded, but fails when it is not.